pyproject.toml
[project]
name = "multimodal-finance-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "chromadb>=0.6.0",
    "deepeval>=3.9.6",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "huggingface-hub[hf-xet]>=1.9.1",
    "langchain>=1.2.15",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.27",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langsmith>=0.7.26",
    "loguru>=0.7.3",
    "mcp[cli]>=1.27.0",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "requests>=2.33.1",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.23",
    "torch>=2.7.0",
    "transformers>=5.5.0",
    "unstructured[pdf]>=0.17.0",
    "uvicorn[standard]>=0.44.0",
    "yfinance>=1.2.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]


[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

# PyTorch CUDA 12.8 wheels — uv fetches torch from this index only
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
